Add files via upload
The discussion on careers underscored the pivotal role played by the Information and Communication Technology (ICT) sector in fostering employment and driving overall economic expansion. The interconnection within the ICT sector enables companies to sustain competitiveness on a global scale and appeal to a broader audience. As students engaged in an ICT career, it is imperative for us to be flexible and committed to perpetual learning.  In essence, I learn that we must instill a mindset emphasizing the requisite skills for our future, such as effective time management, critical thinking, creativity, and more.